requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
jiter==0.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.12.1
networkx==3.4.2
numpy==2.1.2
openai==1.53.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
propcache==0.2.0
protobuf==5.28.3
pyarrow==18.0.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
safetensors==0.4.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
streamlit==1.39.0
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.6
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.17.1
